+++ apps/demo-ui-components/src/app/routes/ui-card.tsx
@@ -1,7 +1,15 @@
 import '@r7/ui-card/css'
 import '@r7/ui-layout/css'
 import { Container, Proportional, SpecialOne } from '@r7/ui-layout'
-import { TwoCardsSquarePhoto, BreakingNews, Card, ProportionalPhoto, Guerra } from '@r7/ui-card'
+import {
+  TwoCardsSquarePhoto,
+  BreakingNews,
+  Card,
+  ProportionalPhoto,
+  Guerra,
+  TextOverPhoto,
+  TitleWithBullets,
+} from '@r7/ui-card'
 import { Bullet } from '@r7/ui-base-components'
 
 export function UiCard() {
@@ -38,6 +46,23 @@ export function UiCard() {
           />
         </Proportional.One>
 
+        <Separator />
+
+        <Proportional.One>
+          <TitleWithBullets
+            title="Leonardo Bricio contrai dengue e desabafa sobre estar doente e sozinho"
+            hat="Economia"
+            renderBullets={
+              <>
+                <Bullet url="#">Câmara aprova MP que adia Lei de Dados para 31 de dezembro</Bullet>
+                <Bullet url="#">Casa Verde e Amarela terá financiamento com juros de 4,25%</Bullet>
+                <Bullet url="#">
+                  Guedes defende fim de deduções do IR para Renda Brasil chegar a R$ 300 aaaa aaaa a
+                </Bullet>
+              </>
+            }
+          />
+        </Proportional.One>
         <Separator />
 
         <Proportional.One>
@@ -97,8 +122,44 @@ export function UiCard() {
             <SpecialOne.Main
               rowOne={
                 <>
-                  <LargeMainBlock />
-                  <LargeMainBlock />
+                  <TextOverPhoto
+                    layout="half"
+                    renderTitle={({ CardTitle }) => (
+                      <CardTitle>
+                        Clientes dão Kombi para ex-galã Daniel Erthal trabalhar em Copacabana
+                      </CardTitle>
+                    )}
+                    renderHat={
+                      <Card.HatWrapper>
+                        <Card.HatTitle color="high">vendedor ambulante</Card.HatTitle>
+                      </Card.HatWrapper>
+                    }
+                    renderImage={
+                      <img
+                        src="//img.r7.com/images/daniel-erthal-03012024135027266?resize=536x326&crop=607x369 34 0&dimensions=536x326"
+                        alt="Clientes dão Kombi para ex-galã Daniel Erthal trabalhar em Copacabana"
+                      />
+                    }
+                  />
+                  <TextOverPhoto
+                    layout="vertical"
+                    renderTitle={({ CardTitle }) => (
+                      <CardTitle>
+                        Leonardo Bricio contrai dengue e desabafa sobre estar doente e sozinho
+                      </CardTitle>
+                    )}
+                    renderHat={
+                      <Card.HatWrapper>
+                        <Card.HatTitle color="high">'Me recuperando'</Card.HatTitle>
+                      </Card.HatWrapper>
+                    }
+                    renderImage={
+                      <img
+                        src="//img.r7.com/images/leonardo-bricio-10012024120558929?resize=208x324&crop=542x844 454 0&dimensions=208x324"
+                        alt="Leonardo Bricio contrai dengue e desabafa sobre estar doente e sozinho"
+                      />
+                    }
+                  />
                 </>
               }
               rowTwo={