+++ apps/demo-ui-components/src/app/routes/ui-layout.tsx
@@ -1,5 +1,14 @@
 import '@r7/ui-layout/css'
-import { Container, SpecialOne, Proportional, Magazine } from '@r7/ui-layout'
+import '@r7/ui-section-heading/css'
+import {
+  LatestNews,
+  Container,
+  SpecialOne,
+  Proportional,
+  Magazine,
+  SectionWrapper,
+} from '@r7/ui-layout'
+import { SectionHeading } from '@r7/ui-section-heading'
 
 export function UiLayout() {
   const LargeMainBlock = () => (
@@ -22,7 +31,7 @@ export function UiLayout() {
 
   return (
     <Container>
-      <div className="flex flex-col gap-xxxs xl:gap-xs">
+      <SectionWrapper>
         <SpecialOne
           renderMain={
             <SpecialOne.Main
@@ -42,6 +51,20 @@ export function UiLayout() {
             />
           }
           renderSidebar={<SidebarBlock />}
+        />
+        <SectionHeading color="#4766ac">
+          <SectionHeading.Title>Últimas</SectionHeading.Title>
+          <SectionHeading.Line />
+        </SectionHeading>
+        <LatestNews
+          renderMain={<LargeMainBlock />}
+          renderSidebar={
+            <LatestNews.Sidebar
+              rowOne={<SidebarBlock />}
+              rowTwo={<SidebarBlock />}
+              rowThree={<SidebarBlock />}
+            />
+          }
         />
 
         <Proportional.One>
@@ -76,7 +99,7 @@ export function UiLayout() {
           <LargeMainBlock />
           <LargeMainBlock />
         </Magazine.Two>
-      </div>
+      </SectionWrapper>
     </Container>
   )
 }