+++ libs/ui-article-delivery/src/components/accessibility-bar/accessibility-bar-list/accessibility-bar-list.tsx
@@ -6,7 +6,7 @@ export type AccessibilityBarListProps = {
 
 export function AccessibilityBarList({ children }: AccessibilityBarListProps) {
   return (
-    <ul aria-label="Barra de acessibilidade" className="article-flex article-list-none">
+    <ul aria-label="Barra de acessibilidade" className="article-flex !article-list-none">
       {children}
     </ul>
   )