+++ libs/ui-layout/src/components/magazine/magazine.tsx
@@ -6,24 +6,24 @@ type MagazineProps = {
 }
 
 function MagazineOne({ children }: MagazineProps) {
-  const blocks = Children.toArray(children)
+  const [block1, block2, block3] = Children.toArray(children)
 
   return (
     <Grid>
-      <div className="lg:layout-col-span-12">{blocks[0]}</div>
-      <div className="lg:layout-col-span-5">{blocks[1]}</div>
-      <div className="lg:layout-col-span-7">{blocks[2]}</div>
+      <div className="md:layout-col-span-24 lg:layout-col-span-12">{block1}</div>
+      <div className="md:layout-col-span-11 lg:layout-col-span-5">{block2}</div>
+      <div className="md:layout-col-span-13 lg:layout-col-span-7">{block3}</div>
     </Grid>
   )
 }
 
 function MagazineTwo({ children }: MagazineProps) {
-  const blocks = Children.toArray(children)
+  const [block1, block2] = Children.toArray(children)
 
   return (
     <Grid>
-      <div className="lg:layout-col-span-17">{blocks[0]}</div>
-      <div className="lg:layout-col-span-7">{blocks[1]}</div>
+      <div className="md:layout-col-span-17">{block1}</div>
+      <div className="md:layout-col-span-7">{block2}</div>
     </Grid>
   )
 }