+++ libs/ui-layout/src/index.ts
@@ -1,5 +1,7 @@
 export * from './styles/styles'
 export * from './components/container'
+export * from './components/section-wrapper'
 export * from './components/special-one'
 export * from './components/proportional'
 export * from './components/magazine'
+export * from './components/latest-news'