+++ libs/ui-layout/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [0.5.0](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-layout-0.4.0...ui-layout-0.5.0) (2024-01-17)
+
+### Features
+
+- Created latest news chain, storybook and cypress ([90f700d](https://gitlab.ir7.com.br/r7/front-monorepo/commit/90f700d8052ec48cd3ea61dbfb57fe7328d3ec7c))
+
+# [0.4.0](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-layout-0.3.1...ui-layout-0.4.0) (2024-01-11)
+
+### Features
+
+- **ui-layout:** creating section wrapper to add gaps between grids ([b761fd6](https://gitlab.ir7.com.br/r7/front-monorepo/commit/b761fd6e27264192e9de85f2c08a0f9fe289f3cf))
+
+## [0.3.1](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-layout-0.3.0...ui-layout-0.3.1) (2024-01-11)
+
+### Features
+
+- **ui-layout:** added height 100% to sidebar of specialOne ([7fa6705](https://gitlab.ir7.com.br/r7/front-monorepo/commit/7fa6705b9ebb67b03d3910780681583352698e69))
+
 # [0.3.0](https://gitlab.ir7.com.br/r7/front-monorepo/compare/ui-layout-0.2.0...ui-layout-0.3.0) (2023-12-15)
 
 ### Features